+++ src/entidades/Carro.java
@@ -1,44 +1,9 @@
 package entidades;
 
-public class Carro {
+public class Carro extends Veiculo {
 
-	private String fabricante;
-	private String modelo;
-	private int ano;
-	private String cor;
-	
 	public Carro(String fabricante, String modelo, int ano, String cor) {
-		super();
-		this.fabricante = fabricante;
-		this.modelo = modelo;
-		this.ano = ano;
-		this.cor = cor;
+		super(fabricante, modelo, ano, cor);
 	}
-	
-	public String getFabricante() {
-		return fabricante;
-	}
-	public void setFabricante(String fabricante) {
-		this.fabricante = fabricante;
-	}
-	public String getModelo() {
-		return modelo;
-	}
-	public void setModelo(String modelo) {
-		this.modelo = modelo;
-	}
-	public int getAno() {
-		return ano;
-	}
-	public void setAno(int ano) {
-		this.ano = ano;
-	}
-	public String getCor() {
-		return cor;
-	}
-	public void setCor(String cor) {
-		this.cor = cor;
-	}
-	
-	
+
 }
